feat: remove 'small' or irrelevant words from related terms search buttons
Add array on the BookDetails component for filtering out small words to related terms links. (e.g. 'the', 'an', 'in', etc.) In the future this could be it's own module, but for simplicity it has been included as a field in the BookDetails component class. Terms are compared against the 'toLowerCase' version of a title's words, so filtering words must be lowercase at this point.