fix: pass more robust function to update book status for BookDetails component for cases where book is not already in user shelf
Add the 'addBookAndUpdate' function to update book status on the BookDetails component instead of the previous 'updateReadingStatus' function. This fixes a bug where adding a none shelf book from the BookDetails view will add it to the users shelf on the API, but not add it to the current app state, requiring a browser refresh to display the book in the user shelf.