feature: display dynamic URL search terms as value in search input initially
Provide UI enhancement by displaying the query term prepopulated in the search input field when a user navigates to a dynamic search route. This adds a step when a new term is to be input, but this will be solved next.  If there is no dynamic term present, the placeholder message will be displayed.