feature: clear search input field on click
Add UX functionality to clear search field on click. This mainly provides experience enhancement for when a user wants to search for a new term, but a dynamically prepopulated term is already the value. There is a risk than accidental clicks will clear a search, but overall it is the desired behavior. Clicking into the empty input value does not trigger a new search, so the search will not be lost in case of an accidental click.